+++ src/main/java/se/experis/repositories/CharacterRepository.java
@@ -3,7 +3,7 @@ package se.experis.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import se.experis.models.Character;
 
-public interface CharacterRepository extends JpaRepository<Character, Integer> {
+public interface CharacterRepository extends JpaRepository<Character, Long> {
 
 
 }
