+++ src/main/java/se/experis/repositories/FranchiseRepository.java
@@ -3,7 +3,7 @@ package se.experis.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import se.experis.models.Franchise;
 
-public interface FranchiseRepository extends JpaRepository<Franchise, Integer> {
+public interface FranchiseRepository extends JpaRepository<Franchise, Long> {
 
 
 }
